requirements.txt
pandas 
numpy 
scikit-learn
matplotlib
seaborn 
shap
streamlit
joblib
pickle
tensorflow
keras

pytest
pyyaml
